Latency computation fixed (final)
The latency computation must explicitly consider
all offsets in the given range although this leads
to worse results, because we may NOT deduce, that
the WCET was derived with the upper bound offset.
This was fixed now.

The 'tryPenalizedAlignment' method now consistently
tries to align the loop / procedure head with the
TDMA borders and checks whether this yields better
results.