Improved precision of alignment-sensitive single-block analysis
Even if a block starts with unknown offsets ([0,max_offset-1]),
we may be able to bound the waiting time for bus accesses which
occur after the first bus access in the block has waited for a
new TDMA slot. After that first access has waited we know the
offset to the TDMA schedule in the worst-case scenario, and can
thus use that fixed offset as long as we have an uninterruptible
instruction sequence (until block end or call).